I overhauled the Mesh/RenderComponent relationship. This allows for Materials to be specified per mesh object. Therefore, any mesh that may be the same mesh but has different materials will be created as a separate mesh. The benefit of this is that any static mesh that uses the same material can be associated with a single mesh, but multiple transformation matrices. Therefore, there is less draw calls required for many static meshes of the same mesh with the same material as well as less space required for storing this information in the program.
The Entity Manager now stores rendering components in a hashmap with the Mesh pointer as a Hashkey. This trick allows many static meshes to maintain their individuality w.r.t. physics and identity/message, but all use the same mesh and rendercomponent as each other for faster rendering speeds.

This change also bred a small change to the scene loader. Particularly for Static Meshes, now you can specify an array of positions along with the material, mesh and shader for the objects and it will generate them in an instanced manner as explained above.