Added Spatial Definitions for all Entities. This is used to generate some coverage across multiple cells in the Spatial Data Map as needed.
Generally, the Spatial information is stored in the mesh. It defaults to returning the Bounding Box dimensions if they exists, otherwise, it will return the computed dimensions  of the mesh which is a basic cubic convex hull.

Added a PLANE_ENTITY enum to the Entity Types that is ignored by the spatial data map.