Added Color Blending for Overlapping Cells in the SpatialDataMap debug mode.
Made SpatialDataMap a singleton.

Added Dynamic Entity handling for Spatial Data Map as well as debug mode drawing for Dynamic Entities. Need to refactor this a bit to make it a bit more accurate.

Will also look as speeding this up a bit more

Also need to add debug wrappers so that the drawing calculations aren't used in Release mode.