Rocket Information can be loaded in via the Scene Loader. It just takes in the mesh, materials, shader, bounding box, etc. information for the Rocket and saves them in the Scene Loader.
Then, when the HovercraftEntity wants to generate a Rocket entity, they call the generation from the Scene Loader which will generate a rocket with the applied Rocket Properties loaded from the scene file.

I think this will be a pattern that will be used for generating dynamic objects such as the players and bots. This allows us to specify how to load them in the scene file, but then we can load a number of them internally as we need to.

Also added a function to the Physics Component to get a direction vector for the current physics object.